% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.R
\name{ace_ui}
\alias{ace_ui}
\alias{ace_server}
\title{UI and server functions to tweak aceEditor so that it works with
  shinyStorePlus}
\usage{
ace_ui(id, value, ..., debug = FALSE)

ace_server(id)
}
\arguments{
\item{id}{character. ID for the module.}

\item{value}{character. Default text.}

\item{...}{other parameters passed to [shinyAce::aceEditor()]}

\item{debug}{logical. If TRUE, the internally mirroring textArea will show up.}
}
\value{
`ace_ui()` returns the UI elements. `ace_server()` returns the
  reactive value holding the content of the aceEditor.
}
\description{
UI function `ace_ui()` and server function `ace_server()` makes
the content written inside [shinyAce::aceEditor()] can be saved in indexeddb.
Internally, the module set up a [shiny::textAreaInput()] to mirror the
content of the Ace Editor.
}
\details{
The internal id for aceEditor is "editor" and the content can be
accessed with `session$userData[[_id_]]`.
}
