% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run the demo app}
\usage{
run_app(debug = FALSE)
}
\arguments{
\item{debug}{logical}
}
\value{
nothing. This function runs the app.
}
\description{
This app demonstrates how
}
\examples{
\dontrun{
  run_app()
}
}
