% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyace4ssp-package.R
\docType{package}
\name{shinyace4ssp-package}
\alias{shinyace4ssp}
\alias{shinyace4ssp-package}
\title{shinyace4ssp: Mirror AceEditor Input to Text Area to Make the Former SSP Compatible}
\description{
This is a demonstration of a hackish work around to make shinyAce's aceEditor input to work with shinyStorePlus (indexeddb).
}
\author{
\strong{Maintainer}: Kenji Sato \email{kenji@kenjisato.jp} (\href{https://orcid.org/0009-0000-4520-2560}{ORCID}) [copyright holder]

}
\keyword{internal}
